started to work on feed page
