+++ JS/auth/constants.mjs
@@ -6,7 +6,9 @@ export const email_address = document.getElementById('email_address');
 export const password = document.getElementById('password');
 export const confirm_password = document.getElementById('confirm_password');
 
-//API base URL
+
+//API URLs
 export const API_BASE_URL = 'https://api.noroff.dev/api/v1';
 export const REGISTER_URL = '/social/auth/register';
 export const LOGIN_URL = '/social/auth/login';
+export const ALL_POSTS_URL = '/social/posts';
