+++ JS/auth/login.mjs
@@ -34,6 +34,9 @@ if (logInForm) {
 
                 const json = await response.json();
                 const accessToken = json.accessToken;
+
+                //The setItem method is used, which takes two arguments:
+                // a key ('accessToken' in this case) and a value (accessToken). 
                 localStorage.setItem('accessToken', accessToken);
                 console.log(json);
                 return json;
@@ -41,9 +44,7 @@ if (logInForm) {
                 console.log(error);
             }
         }
-
         loginUser((API_BASE_URL + LOGIN_URL), userLogin);
-
     })
 }
