+++ JS/feed/feed.mjs
@@ -1,0 +1,24 @@
+import { API_BASE_URL } from "../auth/constants.mjs";
+import { ALL_POSTS_URL } from "../auth/constants.mjs";
+
+async function fetchWithToken(url) {
+    try {
+        const token = localStorage.getItem('accessToken');
+        const getData = {
+            method: 'GET',
+            headers: {
+                'Content-Type': 'application/json',
+                Authorization: `Bearer ${token}`,
+            },
+        };
+        const response = await fetch(url, getData);
+        console.log(response);
+        const json = await response.json();
+        console.log(json);
+
+    } catch (error) {
+        console.log(error);
+    }
+}
+
+fetchWithToken(API_BASE_URL + ALL_POSTS_URL);
