+++ feed/index.html
@@ -9,8 +9,9 @@
     <title>Feed Page || MedSci InsightsMedSci Insights:A platform for researchers to share their research invention
     </title>
     <link rel="stylesheet" href="../node_modules/bootstrap/dist/css/bootstrap.min.css">
-    <script defer src="../node_modules/bootstrap/dist/js/bootstrap.min.js"></script>
-    <script type="module" src="../JS/feed/feed.js"></script>
+    <!--<script defer src="../node_modules/bootstrap/dist/js/bootstrap.min.js"></script>-->
+    <script defer src="../node_modules/bootstrap/dist/js/bootstrap.bundle.min.js"></script>
+    <script type="module" src="../JS/feed/feed.mjs"></script>
     <link rel="stylesheet" href="../dist/css/index.css">
     <link rel="stylesheet" href="../css/customcss.css">
     <!---Fonts links-->
